requirements.txt
# Arus Python Dependencies — Google AI Ecosystem
# Install: pip install -r requirements.txt

# === Agent Framework (Google ADK + Gemini) ===
google-adk==1.27.1
google-genai==1.67.0

# === MCP Protocol ===
mcp==1.26.0
fastmcp==3.1.1

# === Google Cloud (Firestore persistence) ===
google-cloud-firestore==2.19.0

# === Web Framework ===
fastapi==0.135.1
uvicorn==0.42.0
python-multipart==0.0.20

# === Data Validation ===
pydantic==2.12.5

# === Simulation Engine ===
numpy==2.4.3
scipy
pathfinding==1.0.20

# === Configuration ===
pyyaml
python-dotenv

# === Dev / Testing ===
pytest
pytest-asyncio
httpx
httpx-ws
